API Reference

API Reference

HyQMOM.GramianMomentEquations1D — Type

struct GramianMomentEquations1D{Mp1, N, RealT <: Real} <: Trixi.AbstractEquations{1, Mp1}

A struct representing one-dimensional Gramian moment equations for kinetic theory.

closure::Symbol

inv_Kn::RealT: The inverse of the Knudsen number, used in the relaxation source term function GramianMomentEquations1D(M::Integer, Knudsen::Real, closure::Union{Symbol,String}=:ExtGram; χ_set="optimal")

n::Int: Internal parameter derived from the number of moments

extended::Bool: Flag indicating whether to use extended formulation

Constructor

GramianMomentEquations1D(M::Integer, Knudsen::Real, extended=true::Bool)

Constructs a GramianMomentEquations1D system with M moments and specified Knudsen number.

Arguments

M::Integer: Number of moments (must be greater than 1)

Knudsen::Real: Knudsen number for the system

closure::Union{Symbol,String}: Closure type, can be "Gram", "ExtGram", or "Grad"

Notes

The system has M+1 variables in total

The parameter χ is calculated differently depending on whether M is even or odd

HyQMOM.GramianMomentEquations1D3D — Type

struct GramianMomentEquations1D3D{Mp1, N, RealT <: Real} <: Trixi.AbstractEquations{1, Mp1}

A struct representing one-dimensional Gramian moment equations for kinetic theory.

closure::Symbol

inv_Kn::RealT: The inverse of the Knudsen number, used in the relaxation source term function GramianMomentEquations1D(M::Integer, Knudsen::Real, closure::Union{Symbol,String}=:ExtGram; χ_set="optimal")

n::Int: Internal parameter derived from the number of moments

extended::Bool: Flag indicating whether to use extended formulation

Constructor

GramianMomentEquations1D(M::Integer, Knudsen::Real, extended=true::Bool)

Constructs a GramianMomentEquations1D system with M moments and specified Knudsen number.

Arguments

M::Integer: Number of moments (must be greater than 1) - for the moment hard-coded to M=4

Knudsen::Real: Knudsen number for the system

closure::Union{Symbol,String}: Closure type, can be "Gram", "ExtGram", or "Grad"

Notes

The system has M+1 variables in total

The parameter χ is calculated differently depending on whether M is even or odd

HyQMOM.SaveTriangulationCallback — Type

Callback struct that triangulates and writes the 2D tree mesh and the solution variables. Can also be used for 1D Tree meshes, in which case a continuous piecewise linear solution is constructed.

#Configuration:

-solution_variables: The set of solution variables to write (via Trixi solution variables function, such as 'cons2prim')

-interval: Save every interval time-steps. An interval <= 0 means do not save interval-based.

-time_interval: Save at every simulation time multiple of the time interval. A time interval <= 0 disables this functionality. Cannot be used together with interval based saving. The callback guarantees that the interval multiples are hit by reducing the time-step size if needed.

-save_initial_solution, save_final_solution, output_directory

-file_format: supported file-formats are "dat", "szplt", "vtu" "h5". '.dat' is the Tecplot Tabular data file written in ascii, readable by both Tecplot and Paraview. '.szplt' is the Tecplot proprietary format only readable by Tecplot. Use of szplt writing requires the LDLIBRARYPATH to contain the libtecio.so shared library file. Alternatively libtecio.so can also be present in the working directory. '.vtu' is the native Paraview format for an unstructured grid. '.h5' is a dump of the relevant data matrices into a hdf5 file, for more direct reading with e.g. Mathematica. '.tsv' 1D only file format containing the vertex data 'trixi' use Trixi.SaveSolutionCallback to generate the output, instead of this triangulation callback.

-append_solution: If true, write all solution data into one file. Only the dat, szplt, h5 and tsv file formats support the append option. vtu will simply ignore the append flag. Note that Tecplot only recognizes time evolutions when they are in one file, while Paraview is the exact opposite.

-verbose: print a short message to console every time output is written

-clear_out_dir: delete the contents of the output directory at initialization

HyQMOM.closure — Method

closure(u, equations::GramianMomentEquations1D3D)

Decides which closure implementation is used based on the equations.closure value

HyQMOM.closure — Method

closure(u, equations::GramianMomentEquations1D)

Decides which closure implementation is used based on the equations.closure value

HyQMOM.relaxation_source — Method

relaxation_source(u, x, t, equations)

RHS = -1/Kn * (U - U_eq) for the 1D3D 10-moment system.

HyQMOM.zero_source — Method

Zero source, no RHS

Trixi.flux — Method

Trixi.flux(u, orientation::Integer, equations::GramianMomentEquations1D3D{Mp1}) where {Mp1}
Flux function: F(U) = u_{k+1}, k=0,…,M where u_{M+1} = C(u0, u1, …, uM)

Trixi.flux — Method

Trixi.flux(u, orientation::Integer, equations::GramianMomentEquations1D{Mp1}) where {Mp1}

Flux function: F(U) = u_{k+1}, k=0,…,M where u_{M+1} = C(u0, u1, …, uM)